Project Descriptor
For the purposes of our neural network, we will require a system that involves the processing of pixel data. A convolutional neural network is specifically used in image recognition and tasked in the identifying & recognizing of objects. In this application, we will be using a Convolutional Neural Network to analyze & identify guns using a database of 500,000 images collected from a research study specializing in image recognition of weapons. 40,000 images were used in the training of our neural network, which enabled our development of a CNN that would be able to recognize if an object was or wasn’t a gun.